+++ src/gmail/auth.ts
@@ -6,119 +6,80 @@ import { SCOPES_GMAIL } from './constant';
 import { ICredentials } from './types';
 import { getGoogleConfigs } from './util';
 
-const gmail: any = google.gmail('v1');
+const gmail: any = google.gmail({
+  version: 'v1',
+});
 
 export const gmailClient = gmail.users;
 
-const getOauthClient = async () => {
+export const getOauthClient = async () => {
   const { GOOGLE_CLIENT_ID, GOOGLE_CLIENT_SECRET } = await getGoogleConfigs();
 
   const GMAIL_REDIRECT_URL = `${getEnv({ name: 'DOMAIN' })}/gmaillogin`;
 
   if (!GOOGLE_CLIENT_ID || !GOOGLE_CLIENT_SECRET) {
-    return debugGmail(`
+    throw new Error(`
       Error Google: Missing env values
       GOOGLE_CLIENT_ID: ${GOOGLE_CLIENT_ID}
       GOOGLE_CLIENT_SECRET: ${GOOGLE_CLIENT_SECRET}
     `);
   }
 
-  let response;
-
   try {
-    response = new google.auth.OAuth2(GOOGLE_CLIENT_ID, GOOGLE_CLIENT_SECRET, GMAIL_REDIRECT_URL);
+    return new google.auth.OAuth2(GOOGLE_CLIENT_ID, GOOGLE_CLIENT_SECRET, GMAIL_REDIRECT_URL);
   } catch (e) {
     debugGmail(`
       Error Google: Could not create OAuth2 Client with
       ${GOOGLE_CLIENT_ID}
       ${GOOGLE_CLIENT_SECRET}
     `);
-    return;
+
+    throw e;
   }
-
-  return response;
 };
 
-/**
- * Get OAuth client with given credentials
- */
-export const getAuth = async (credentials: ICredentials, accountId?: string) => {
-  // Google OAuthClient ================
-  const oauth2Client = await getOauthClient();
-
-  oauth2Client.on('tokens', async tokens => {
-    await refreshAccessToken(accountId, credentials);
-
-    credentials = tokens;
-  });
-
-  oauth2Client.setCredentials(credentials);
-
-  return oauth2Client;
-};
-
-/**
- * Get auth url depends on google services such us gmail, calendar
- */
 export const getAuthorizeUrl = async (): Promise<string> => {
-  // Google OAuthClient ================
-  const oauth2Client = await getOauthClient();
-
-  const options = { access_type: 'offline', scope: SCOPES_GMAIL };
-
-  let authUrl;
-
-  debugGmail(`
-    Google OAuthClient generate auth url with following data
-    ${options}
-  `);
+  debugGmail(`Google OAuthClient generate auth url`);
 
   try {
-    authUrl = oauth2Client.generateAuthUrl(options);
+    const auth = await getOauthClient();
+
+    return auth.generateAuthUrl({ access_type: 'offline', scope: SCOPES_GMAIL });
   } catch (e) {
     debugGmail(`Google OAuthClient failed to generate auth url`);
+    throw e;
   }
-
-  return authUrl;
 };
 
-/**
- * Get access token from gmail callback
- */
-export const getAccessToken = async (code: string) => {
-  let accessToken;
-
+export const getAccessToken = async (code: string): Promise<ICredentials> => {
   debugGmail(`Google OAuthClient request to get token with ${code}`);
 
-  // Google OAuthClient ================
-  const oauth2Client = await getOauthClient();
-
   try {
-    accessToken = await new Promise((resolve, reject) =>
+    const oauth2Client = await getOauthClient();
+
+    return new Promise((resolve, reject) =>
       oauth2Client.getToken(code, (err: any, token: ICredentials) => {
         if (err) {
-          return reject(err.response.data.error);
+          return reject(new Error(err.response.data.error));
         }
+
+        // set access token
+        oauth2Client.setCredentials(token);
 
         return resolve(token);
       }),
     );
   } catch (e) {
     debugGmail(`Error Google: Google OAuthClient failed to get access token with ${code}`);
+    throw e;
   }
-
-  return accessToken;
 };
 
-/**
- * Refresh token and save when access_token expires
- */
 export const refreshAccessToken = async (_id: string, tokens: ICredentials): Promise<void> => {
   const account = await Accounts.findOne({ _id });
 
   if (!account) {
-    debugGmail(`Error Google: Account not found id with ${_id}`);
-    return;
+    return debugGmail(`Error Google: Account not found id with ${_id}`);
   }
 
   account.token = tokens.access_token;
@@ -128,7 +89,7 @@ export const refreshAccessToken = async (_id: string, tokens: ICredentials): Pro
   }
 
   if (tokens.expiry_date) {
-    account.expireDate = tokens.expiry_date;
+    account.expireDate = tokens.expiry_date.toString();
   }
 
   await account.save();