+++ src/index.ts
@@ -18,10 +18,6 @@ import initTwitter from './twitter/controller';
 import initDaily from './videoCall/controller';
 import initWhatsapp from './whatsapp/controller';
 
-initRedis();
-
-initConsumer();
-
 const app = express();
 
 const rawBodySaver = (req, _res, buf, encoding) => {
@@ -133,7 +129,10 @@ app.use((error, _req, res, _next) => {
 const { PORT } = process.env;
 
 app.listen(PORT, () => {
-  connect().then(() => {
+  connect().then(async () => {
+    await initRedis();
+    await initConsumer();
+
     // Initialize startup
     init();
   });